手牌数字段类型修改

+++ src/main/java/com/dx/facade/texas/resp/DxGameBillResp.java
@@ -88,5 +88,5 @@ public class DxGameBillResp implements Serializable {
     /**
      * Walks手牌数
      */
-    private Long walksHandNumber;
+    private Integer walksHandNumber;
 }
